refactor: extract sqlite backend repository wiring into a mixin (#2050)
